# sso_manual/openapi.json
{
  "swagger": "2.0",
  "definitions": {
...
    "RequestOauth2Introspect": {
      "type": "object",
      "properties": {
        "token": {
          "type": "string"
        }
      },
      "required": [
        "token"
      ]
    },
    "RequestOauth2Token": {
      "type": "object",
      "properties": {
        "code": {
          "type": "string"
        },
        "grant_type": {
          "type": "string"
        },
        "redirect_uri": {
          "type": "string"
        },
        "refresh_token": {
          "type": "string"
        }
      },
      "required": [
        "grant_type"
      ]
    },
...
  },
  "paths": {
...
    "/v2/auth/delete": {
      "post": {
        "summary": "Delete user interface",
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "parameters": [
          {
            "in": "query",
            "name": "client_id",
            "type": "string",
            "format": "uuid"
          },
          {
            "in": "query",
            "name": "code",
            "type": "string"
          },
          {
            "in": "query",
            "name": "redirect_uri",
            "type": "string"
          },
          {
            "in": "query",
            "name": "response_type",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "csrf_token",
            "required": true,
            "type": "string"
          },
          {
            "in": "formData",
            "name": "password",
            "required": true,
            "type": "string"
          }
        ]
      }
    },
...
    "/v2/auth/password-reset": {
      "post": {
        "summary": "User reset password interface",
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "parameters": [
          {
            "in": "query",
            "name": "code",
            "required": true,
            "type": "string"
          },
          {
            "in": "query",
            "name": "response_type",
            "required": true,
            "type": "string"
          },
          {
            "in": "formData",
            "name": "csrf_token",
            "required": true,
            "type": "string"
          },
          {
            "in": "formData",
            "name": "password_confirm",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "password_new",
            "type": "string"
          }
        ]
      }
    },
    "/v2/auth/password-update": {
      "post": {
        "summary": "User update password interface",
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "parameters": [
          {
            "in": "query",
            "name": "client_id",
            "required": true,
            "type": "string",
            "format": "uuid"
          },
          {
            "in": "query",
            "name": "message",
            "type": "string"
          },
          {
            "in": "query",
            "name": "redirect_uri",
            "required": true,
            "type": "string"
          },
          {
            "in": "formData",
            "name": "csrf_token",
            "required": true,
            "type": "string"
          },
          {
            "in": "formData",
            "name": "password",
            "required": true,
            "type": "string"
          },
          {
            "in": "formData",
            "name": "password_confirm",
            "required": true,
            "type": "string"
          },
          {
            "in": "formData",
            "name": "password_new",
            "required": true,
            "type": "string"
          }
        ]
      }
    },
    "/v2/auth/register": {
      "post": {
        "summary": "User registration interface",
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "parameters": [
          {
            "in": "query",
            "name": "code",
            "required": true,
            "type": "string"
          },
          {
            "in": "query",
            "name": "response_type",
            "required": true,
            "type": "string"
          },
          {
            "in": "formData",
            "name": "csrf_token",
            "required": true,
            "type": "string"
          },
          {
            "in": "formData",
            "name": "name",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "oauth2_provider",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "password",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "password_allow_reset",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "password_confirm",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "register_type",
            "required": true,
            "type": "string"
          }
        ]
      }
    },
...
    "/v2/oauth2/authorize": {
      "post": {
        "summary": "OAuth2 authorization endpoint",
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "parameters": [
          {
            "in": "query",
            "name": "client_id",
            "required": true,
            "type": "string"
          },
          {
            "in": "query",
            "name": "redirect_uri",
            "required": true,
            "type": "string"
          },
          {
            "in": "query",
            "name": "response_type",
            "required": true,
            "type": "string"
          },
          {
            "in": "query",
            "name": "scope",
            "type": "string"
          },
          {
            "in": "query",
            "name": "state",
            "required": true,
            "type": "string"
          },
          {
            "in": "formData",
            "name": "auth_type",
            "required": true,
            "type": "string"
          },
          {
            "in": "formData",
            "name": "csrf_token",
            "required": true,
            "type": "string"
          },
          {
            "in": "formData",
            "name": "email",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "oauth2_provider",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "password",
            "type": "string"
          }
        ]
      }
    },
    "/v2/oauth2/introspect": {
      "post": {
        "summary": "OAuth2 token introspection endpoint",
        "security": [
          {
            "basicAuth": []
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/RequestOauth2Introspect"
            }
          }
        ]
      }
    },
    "/v2/oauth2/redirect": {
      "get": {
        "summary": "OAuth2 redirect endpoint",
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "parameters": [
          {
            "in": "query",
            "name": "code",
            "required": true,
            "type": "string"
          },
          {
            "in": "query",
            "name": "state",
            "required": true,
            "type": "string"
          }
        ]
      }
    },
    "/v2/oauth2/token": {
      "post": {
        "summary": "OAuth2 token endpoint",
        "security": [
          {
            "basicAuth": []
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/RequestOauth2Token"
            }
          }
        ]
      }
    },
...
  },
  "securityDefinitions": {
    "basicAuth": {
      "type": "basic"
    }
  },
  "info": {
    "version": "v2",
    "title": "Single Sign-On API"
  }
}